tests: move tests closer to sources instead of a separated folder
