+++ lib/helpers/deprecate.test.js
