+++ lib/helpers/match.test.js
@@ -1,4 +1,4 @@
-const match = require('../../lib/helpers/match')
+const match = require('./match')
 
 test('matches all dependencies', async () => {
   const deps = await match({
