+++ lib/report.test.js
@@ -1,5 +1,5 @@
 // Ours
-const report = require('../lib/report')
+const report = require('./report')
 
 // Globals
 let github, repo, sha
