+++ lib/run.test.js
@@ -1,8 +1,8 @@
-jest.mock('../lib/report')
+jest.mock('./report')
 
 // Ours
-const run = require('../lib/run')
-const report = require('../lib/report')
+const run = require('./run')
+const report = require('./report')
 
 // Globals
 let github, repo, sha, deps
