+++ lib/toggle.js
@@ -1,12 +1,16 @@
 // Ours
 const run = require('./run')
 const match = require('./helpers/match')
+// const migrate = require('./helpers/migrate')
 
 async function toggle (context) {
   // 1. Extract necessary info
   const issue = context.issue()
   const { payload, github } = context
   const { sha } = payload.pull_request.head
+
+  // // 2. Migrate if necessary
+  // const body = await migrate(context)
 
   // 2. Match issue patterns
   const deps = await match(payload.pull_request)
