+++ lib/toggle.test.js
@@ -1,4 +1,4 @@
-jest.mock('../lib/run')
+jest.mock('./run')
 
 // Packages
 const { createRobot } = require('probot')
@@ -6,8 +6,8 @@ const { createRobot } = require('probot')
 // Ours
 const app = require('../index')
 const events = require('./events')
-const toggle = require('../lib/toggle')
-const run = require('../lib/run')
+const toggle = require('./toggle')
+const run = require('./run')
 
 // Globals
 let robot, github
