+++ lib/update.test.js
@@ -1,4 +1,4 @@
-jest.mock('../lib/run')
+jest.mock('./run')
 
 // Packages
 const { createRobot } = require('probot')
@@ -6,7 +6,7 @@ const { createRobot } = require('probot')
 // Ours
 const app = require('../index')
 const events = require('./events')
-const run = require('../lib/run')
+const run = require('./run')
 
 // Globals
 let robot, github, repo, sha, deps
